Nested-loops commit

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,8 +6,8 @@
 int main(void)
 {
 	int a = 0;
-	int b = 1;
-	int c = 2;
+	long int b = 1;
+	long int c = 2;
 
 	while (a < 50)
 	{
